+++ demo/NeuralCA.js
@@ -7,14 +7,18 @@
 // https://arxiv.org/abs/2111.13545
 class NeuralCA {
     static Tags = ['2d', 'ca'];
+    constructor() {
+      this.W = new Float32Array([
+        -67,1,2,44,-13,-59,4,30,-1,16,-57,9,-10,-4,-2,-41,    19,-18,-1,8,-4,35,8,0,-4,-4,-1,0,34,31,21,-25,
+        4,13,18,-57,-79,-22,-25,71,-12,-11,24,27,-17,-8,-7,6, 11,10,4,0,4,1,2,7,-26,-33,-15,-3,22,27,20,-34]);
+      this.b = new Float32Array([2,-5,-14,9]);
+    }
     frame(glsl) {
-        const state = glsl({FP:`
-        vec4 rule2(vec4 s, vec4 p) {
-          return 1e-3*(vec4(4,-10,-27,18)/2.+
-            mat4(-67,1,2,44,-13,-59,4,30,-1,16,-57,9,-10,-4,-2,-41)*s+
-            mat4(19,-18,-1,8,-4,35,8,0,-4,-4,-1,0,34,31,21,-25)*p+
-            mat4(4,13,18,-57,-79,-22,-25,71,-12,-11,24,27,-17,-8,-7,6)*abs(s)+
-            mat4(11,10,4,0,4,1,2,7,-26,-33,-15,-3,22,27,20,-34)*abs(p));     
+        const state = glsl({W:this.W, b:this.b, FP:`
+        uniform mat4 W[4];
+        uniform vec4 b;
+        vec4 rule(vec4 s, vec4 p) {
+          return 1e-3*(b + W[0]*s + W[1]*p + W[2]*abs(s) + W[3]*abs(p));     
         }
         void fragment() {
           vec4 s = Src(UV);
@@ -31,7 +35,7 @@ class NeuralCA {
           vec4 p = R(l,u)*vec4(1,1,-1, 1) + R(x,u)*vec4(2,2,0, 2) + R(r,u)*vec4(1,1,1, 1)
                  + R(l,y)*vec4(2,2,-2, 0) +  s*vec4(-12,-12,0, 0) + R(r,y)*vec4(2,2,2, 0)
                  + R(l,d)*vec4(1,1,-1,-1) + R(x,d)*vec4(2,2,0,-2) + R(r,d)*vec4(1,1,1,-1);
-          vec4 ds = rule2(s-0.5, p);  // NCA rule application
+          vec4 ds = rule(s-0.5, p);  // NCA rule application
           FOut = s+ds;
         }`}, {story:2, scale:1/4, filter:'nearest', tag:'state'});
         glsl({tex:state[0], FP:`tex(UV)*2.-.5`});